+++ packages/api-client/src/index.ts
@@ -3,7 +3,5 @@ export {
   USER_TOKEN_STORAGE_KEY,
   USER_REFRESH_TOKEN_STORAGE_KEY,
   FetchOptions,
-  FetchResponseTypes,
 } from './api-client'
 export { ACCESS_TOKEN_STRING, getAccessTokenFromUrl, removeAccessTokenFromUrl } from './utils/url'
-export type { UserData, UploadResponse } from './types'
