+++ packages/api-types/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @globalfishingwatch/api-types
