+++ packages/api-types/README.md
@@ -1,0 +1,9 @@
+# GFW API TYPES
+
+Typescript definitions for [API responses](https://gateway.api.dev.globalfishingwatch.org/swagger#)
+
+Just se them or build it to publish in npm
+
+```bash
+yarn build
+```
