+++ packages/api-types/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@globalfishingwatch/api-types",
+  "version": "1.0.0",
+  "description": "Typescript definitions for API responses",
+  "author": "satellitestudio <contact@satellitestud.io>",
+  "homepage": "https://github.com/GlobalFishingWatch/frontend#readme",
+  "license": "ISC",
+  "main": "src/index.ts",
+  "typings": "src/index.ts",
+  "directories": {
+    "src": "src",
+    "dist": "dist"
+  },
+  "files": ["dist"],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/GlobalFishingWatch/frontend.git"
+  },
+  "scripts": {
+    "start": "rollup -c ./rollup.config.js -w",
+    "build": "yarn clean && cross-env NODE_ENV=production rollup -c rollup.config.js",
+    "clean": "rimraf dist/*",
+    "prepare": "yarn build"
+  },
+  "dependencies": {
+    "@globalfishingwatch/layer-composer": "*"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
