+++ packages/api-types/src/areas-of-interest.ts
@@ -1,0 +1,9 @@
+import type { GeoJSON } from 'geojson'
+
+export interface AOI {
+  id: number
+  name: string
+  area: number
+  geometry?: GeoJSON
+  bbox: number[]
+}
