+++ packages/api-types/src/index.ts
@@ -1,0 +1,6 @@
+export * from './areas-of-interest'
+export * from './datasets'
+export * from './dataviews'
+export * from './resources'
+export * from './user'
+export * from './workspaces'
