+++ packages/api-types/src/user.ts
@@ -1,4 +1,4 @@
-interface UserPermission {
+export interface UserPermission {
   type: 'application' | 'dataset'
   value: 'map-client' | 'data-portal' | 'carrier-portal' | 'indonesia:*' | 'carriers:*' | 'public'
   action: string
@@ -12,9 +12,4 @@ export interface UserData {
   photo: string
   type: string
   permissions: [UserPermission]
-}
-
-export interface UploadResponse {
-  path: string
-  url: string
 }
