+++ packages/dataviews-client/src/index.ts
@@ -1,4 +1,3 @@
 export { default } from './dataviews-client'
 export { default as resolveDataviews } from './resolve-dataviews'
 export { default as resolveEndpoint } from './resolve-endpoint'
-export * from './types'
