+++ packages/react-hooks/src/use-map-interaction/use-map-interaction.ts
@@ -1,8 +1,8 @@
-import debounce from 'lodash/debounce'
-import { isArray } from 'lodash'
 import { useCallback, useEffect, useRef } from 'react'
-import type { Map, MapboxGeoJSONFeature } from '@globalfishingwatch/mapbox-gl'
+import debounce from 'lodash/debounce'
+import isArray from 'lodash/isArray'
 import { Generators } from '@globalfishingwatch/layer-composer'
+import type { Map, MapboxGeoJSONFeature } from '@globalfishingwatch/mapbox-gl'
 import { ExtendedFeature, InteractionEventCallback, InteractionEvent } from '.'
 
 type FeatureStateSource = { source: string; sourceLayer: string }
