+++ packages/react-hooks/src/use-workspace/use-workspace.ts
@@ -1,5 +1,6 @@
 import { useMemo } from 'react'
-import { Dataview, DataviewInstance, resolveDataviews } from '@globalfishingwatch/dataviews-client'
+import { resolveDataviews } from '@globalfishingwatch/dataviews-client'
+import { Dataview, DataviewInstance } from '@globalfishingwatch/api-types'
 
 /**
  * Gets list of dataviews and those present in the workspace, and applies any view or datasetParams from it (merges dataview.defaultView with dataview.view and workspace's dataview.view).
